pyproject.toml
[project]
name = "hackathon-gazprom-2026"
version = "0.1.0"
description = "PDF → Markdown: DocLayout-YOLO + Docling + olmOCR-2-7B"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    # Layout router (YOLOv10)
    "doclayout-yolo>=0.0.4",
    "ultralytics>=8.4.38",
    # Docling — движок для текста/таблиц (Fast Track)
    "docling>=2.0.0",
    "docling-core>=2.0.0",
    "easyocr>=1.7.0",
    # olmOCR-2-7B-1025 — Qwen2.5-VL для crop-fallback
    "olmocr>=0.4.0",
    "transformers>=4.49.0",
    "tokenizers>=0.20.3",
    "accelerate>=0.26.0",
    # Torch
    "torch>=2.5.1",
    "torchvision>=0.20.1",
    # PDF и изображения
    "pymupdf>=1.24.0",
    "pillow>=10.0.0",
    "opencv-python>=4.9.0",
    "numpy>=1.26.0",
    # HuggingFace Hub
    "huggingface-hub>=0.24.0",
]

# ---------------------------------------------------------------------------
# Установка:
#
# --- CUDA (4090, Colab, сервер) ---
#   conda create -n pipeline python=3.12 -y && conda activate pipeline
#   pip install torch torchvision --index-url https://download.pytorch.org/whl/cu124
#   pip install -e .
#   pip install flash-attn --no-build-isolation      # опционально
#   python src/pipeline.py --all
#
# --- Mac (без olmOCR) ---
#   conda create -n pipeline python=3.12 -y && conda activate pipeline
#   pip install torch torchvision --index-url https://download.pytorch.org/whl/cpu
#   pip install -e .
#   python src/pipeline.py --pdf data/raw/document_001.pdf --no-vlm
# ---------------------------------------------------------------------------
